Reorganize types into own modules
