+++ src/rtree.rs
@@ -1,6 +1,6 @@
 mod bounding_box;
 mod dimension_type;
 mod extent;
-mod linear_cost_split;
 mod nodes;
 mod rtree;
+mod splitting_strategies;
